Added output existence condition, avoiding a issue when re-pushing the OpenShift app repository.

+++ openshift/tpb2rss-task.sh
@@ -12,7 +12,7 @@ for n in $( seq `wc -l <  "$FILE"` ); do
 		wget --quiet -O "$DIR/$TMP" "https://thepiratebay.se/search/$ITEM/0/3/0"
 		sed -i -e 's/<a href="[a-z,A-Z,0-9,:,.,\/]*" title="Proxy" target="_blank">Proxy<\/a> \|//g' "$DIR/$TMP"
 		sed -i -e "$( expr `grep -i -n footer -m 1 "$DIR/$TMP" | cut -f1 -d":"` - 1 ),$( expr `wc -l < "$DIR/$TMP"` - 1 )d" "$DIR/$TMP"
-		if [ $( test -f "$DIR/$FILENAME" && echo 1 ) ]; then
+		if [ $( test -f "$DIR/$FILENAME" && echo 1 ) ] && [ $( test -f "$OUTPUTDIR/$FILENAME.xml" && echo 1 ) ]; then
 			if [ "$( cat "$DIR/$TMP" )" != "$( cat "$DIR/$FILENAME" )" ]; then
 				echo "$FILENAME updated"
 				mv "$DIR/$TMP" "$DIR/$FILENAME"
@@ -21,9 +21,9 @@ for n in $( seq `wc -l <  "$FILE"` ); do
 				echo "No updates for $FILENAME"; rm "$DIR/$TMP"
 			fi
 		else
-			echo "$FILENAME created"
-			mv "$DIR/$TMP" "$DIR/$FILENAME"
-			"$OPENSHIFT_REPO_DIR/tpb2rss.py" "$DIR/$FILENAME" "$OUTPUTDIR/$FILENAME.xml"
+			mv "$DIR/$TMP" "$DIR/$FILENAME" &&
+			"$OPENSHIFT_REPO_DIR/tpb2rss.py" "$DIR/$FILENAME" "$OUTPUTDIR/$FILENAME.xml" &&
+			echo "Created feed for $FILENAME"
 		fi
 	fi
 done
